Add DRI2 support
We implement dri auth and gem flink and open which are needed for DRI2
drivers. In this case X creates the color buffers and we can't change
the tiling to linear (since we don't know the stride, which we need to
pass to set_tiling). We'll just render linear data to a tiled buffer for
now.