Implement GTT maps
GTT maps are tricky since they present a linear view of tiled
buffers. We implement this using a shadow buffer where the driver can
access the contents in a linear fashion. At execbuffer time, we copy and
tile the contents of the shadow buffer into the real buffer. This works
for the basic texture upload case, but wont work if the driver tries to
read back data produced by ksim.