Fix MATH hack
It's not enough to make sure ymm0 is first in line to get picked, we
also have to make sure no other AVX2 register holds the value of src0 so
we try to reuse that register.  Since we have to invalidate after
calling the math helper, just invalidate up front.