Quiet uninitialized variable warning
Since gt.vs.enable is a global variable, gcc doesn't reuse the value
in the second if (gt.vs.enable) statement and can't assume that the
first if statement has been run even if we get into the second. Just
refactor the code a bit, which turns out to be a little easier to read
anyway.