Fix another release build warning
Similar to the gt.vs.enable case, gcc can't rely on the assert on
execbuffer2->buffer_count > 0 to guarantee we'll get into the loop,
since the value could change. If we copy it to a local variable, gcc
can know, but we'll refactor this a bit to avoid reusing the bo
variable anyway.